requirements.txt
requests
beautifulsoup4
lxml
pandas
tqdm
python-dotenv
scikit-learn
tenacity
